test(racetest): racetest내 cars 관련 테스트 CarsTest로 이동
Race 내 자동차 리스트 관리 로직 Cars 클래스로 이동하여 racetest 내 cars 관련 테스트를 CarsTest로 이동
