+++ src/test/kotlin/racingcar/RaceTest.kt
@@ -11,69 +11,6 @@ import racingcar.controller.Race
 class RaceTest {
     private val race = Race()
 
-    @DisplayName("자동차를 입력하지 않았을 때")
-    @Test
-    fun inputNone() {
-        val exception =
-            assertThrows<IllegalArgumentException> {
-                race.validateNames(emptyList())
-            }
-
-        assertEquals(
-            ERROR_REQUEST_INPUT,
-            exception.message,
-        )
-    }
-
-    @DisplayName("자동차 수가 두 대 미만일 때")
-    @Test
-    fun inputLessThanTwo() {
-        val names = listOf("lisa")
-        val exception =
-            assertThrows<IllegalArgumentException> {
-                race.validateNames(names)
-            }
-
-        assertEquals(
-            ERROR_CAR_LESS_THAN_TWO,
-            exception.message,
-        )
-    }
-
-    @DisplayName("자동차 수가 20대 초과일 때")
-    @Test
-    fun inputGreaterThanTwenty() {
-        val names =
-            listOf(
-                "crong", "lisa", "hena", "pobi", "poro", "lufy", "joro", "sang", "frank", "usop",
-                "nami", "zzang", "huni", "chul", "yuri", "meang", "chae", "miri", "won", "shin", "bong", "siro",
-            )
-        val exception =
-            assertThrows<IllegalArgumentException> {
-                race.validateNames(names)
-            }
-
-        assertEquals(
-            ERROR_CAR_GREATER_THAN_TWENTY,
-            exception.message,
-        )
-    }
-
-    @DisplayName("중복된 자동차 이름이 있을 때")
-    @Test
-    fun inputDuplicatedName() {
-        val names = listOf("crong", "crong")
-        val exception =
-            assertThrows<IllegalArgumentException> {
-                race.validateNames(names)
-            }
-
-        assertEquals(
-            ERROR_NAME_DUPLICATION,
-            exception.message,
-        )
-    }
-
     @DisplayName("시도 횟수가 int 범위를 초과했을 때")
     @Test
     fun exceedIntRangeRoundNumber() {
@@ -129,10 +66,6 @@ class RaceTest {
     }
 
     companion object {
-        const val ERROR_REQUEST_INPUT = "[Error] 값을 입력해 주세요"
-        const val ERROR_CAR_LESS_THAN_TWO = "[Error] 자동차 이름은 2대 이상 입력해 주세요."
-        const val ERROR_CAR_GREATER_THAN_TWENTY = "[Error] 자동차 이름은 20대 이하로 입력해 주세요."
-        const val ERROR_NAME_DUPLICATION = "[Error] 중복된 자동차 이름은 작성할 수 없습니다."
         const val ERROR_ROUND_LESS_THAN_ONE = "[Error] 1회 이상 입력해 주세요."
         const val ERROR_INPUT_NUMBER_ONLY = "[Error] 숫자만 입력해 주세요."
     }
